完善 CompositeConfigEditor
