+++ Editor/CustomEditor/CompositeConfigEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90e118c8c794a124db14422a858947de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
