+++ Runtime/Attribute/CompositeComponentAttribute.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c4675bd0fb8270c4e9d3ad341f6b02ec
+guid: bde2cd5a410f1c64489e094b9455f3fc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
